+++ changelog.md
@@ -18,6 +18,7 @@ While all _Major_ releases should be reviewed, our only _large_ releases are **v
 - **[fix]** - support unTag property in sweep
 - **[change]** - keep sentence caches, when still valid
 - **[change]** - alias nlp.compile() to .buildTrie()
+- **[fix]** - tagging fixes
 
 #### 14.3.1 [June 2021]
 - **[fix]** - missed caches in .sweep()